% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qread.R
\name{qload}
\alias{qload}
\title{Quartz Load
Reads the object from the data directory into the global namespace using fst.}
\usage{
qload(x)
}
\arguments{
\item{x}{The name of the object to load}
}
\value{
void
}
\description{
Quartz Load
Reads the object from the data directory into the global namespace using fst.
}
