% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qread.R
\name{qread}
\alias{qread}
\title{Quartz Read
Reads the data frame from the data url using read_csv with caching.}
\usage{
qread(
  file,
  col_names = TRUE,
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  quote = "\\"",
  comment = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  name_repair = "unique",
  num_threads = readr_threads(),
  progress = show_progress(),
  show_col_types = should_show_types(),
  skip_empty_rows = TRUE,
  lazy = should_read_lazy()
)
}
\arguments{
\item{url}{The url to read from}
}
\value{
The data frame
}
\description{
Quartz Read
Reads the data frame from the data url using read_csv with caching.
}
