% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qread.R
\name{qsave}
\alias{qsave}
\title{Quartz Save
Writes the data frame to the data directory using fst.}
\usage{
qsave(x, name = as.character(substitute(x)))
}
\arguments{
\item{x}{The object to write}
}
\description{
Quartz Save
Writes the data frame to the data directory using fst.
}
